@@ -71,15 +71,18 @@
     "@storybook/addons": "^5.3.18",
     "@storybook/preset-typescript": "^3.0.0",
     "@storybook/react": "^5.3.18",
-    "@testing-library/jest-dom": "^5.5.0",
-    "@testing-library/react": "^10.0.2",
-    "@testing-library/user-event": "^10.1.0",
+    "@testing-library/dom": "^9.2.0",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "12.1.5",
+    "@testing-library/user-event": "13",
+    "@types/jest": "^29.5.1",
     "@types/jest-axe": "^3.2.2",
     "@types/node": "^14.0.5",
     "@types/react": "^16.9.11",
     "@types/react-dom": "^16.9.3",
-    "@typescript-eslint/eslint-plugin": "^3.0.2",
-    "@typescript-eslint/parser": "^3.0.2",
+    "@types/testing-library__jest-dom": "^5.14.5",
+    "@typescript-eslint/eslint-plugin": "^5.59.1",
+    "@typescript-eslint/parser": "^5.59.1",
     "@zeit/next-css": "^1.0.1",
     "autoprefixer": "^9.7.1",
     "babel-loader": "^8.1.0",
@@ -98,12 +101,14 @@
     "fork-ts-checker-webpack-plugin": "^4.1.3",
     "husky": "^4.2.5",
     "is-ci-cli": "^2.0.0",
-    "jest": "^26.0.1",
-    "jest-axe": "^3.4.0",
+    "jest": "^29.5.0",
+    "jest-axe": "^7.0.1",
+    "jest-config": "^29.5.0",
+    "jest-environment-jsdom": "^29.5.0",
     "jest-next-dynamic": "^1.0.1",
-    "jest-runner-eslint": "^0.9.0",
+    "jest-runner-eslint": "^2.0.0",
     "jest-watch-select-projects": "^2.0.0",
-    "jest-watch-typeahead": "^0.6.0",
+    "jest-watch-typeahead": "^2.2.2",
     "lint-staged": "^10.0.0-1",
     "module-alias": "^2.2.2",
     "netlify-cli": "^2.20.2",
@@ -124,7 +129,7 @@
     "tailwindcss-animations": "^2.0.0",
     "tailwindcss-pseudo-elements": "^1.1.2",
     "ts-loader": "^7.0.2",
-    "typescript": "^3.7.2",
+    "typescript": "^5.0.4",
     "webpack": "^4.41.2"
   }
 }
