@@ -1,12 +1,12 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-const fetch = (...args) => import('node-fetch').then(({default: fetch}) => fetch(...args));
-const sanityClient = require('@sanity/client');
+import fetch from 'node-fetch';
+import sanityClient from '@sanity/client';
+import {config as dotEnvConfig} from 'dotenv';
 
 // Read env variables.
-require('dotenv').config();
+dotEnvConfig();
 const { NEWSLETTER_API_KEY, NEWSLETTER_SUBSCRIBERS_GROUP_ID, SLACK_WEBHOOK_URL } = process.env;
 
-exports.handler = async (event) => {
+export async function handler(event) {
   const { payload, site } = JSON.parse(event.body);
 
   console.log(`NEW SUBMISSION FOR ${payload.form_name}:`, payload.data);
